Bug Fix
When entering text in Change Shader Effect tab there were some issues.  Entering a decimal point  or comma would cause the parse code to run immediately and parse the number
to an incomplete value.  For example entering 1,5 would result in 15 being entered instead of the intended 1,500.  This issue is resolve.  This should work correctly in all cultures, though it has only been tested in Italian.

git-svn-id: https://shazzam.svn.codeplex.com/svn@40303 2327b42d-5241-43d6-9e2a-de5ac946f064